# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend CMS API untuk SGS Law Firm.\n\nAdmin API menggunakan JWT Bearer Token.\nSebagian besar admin endpoint juga membutuhkan X-Timestamp, X-Nonce, dan X-Signature.\nUpload gambar hanya membutuhkan JWT karena menggunakan multipart/form-data.\n\nFormat canonical X-Signature:\nMETHOD + newline + REQUEST_URI + newline + TIMESTAMP + newline + NONCE + newline + SHA256_BODY",
        "title": "SGS CMS API",
        "contact": {
            "name": "SGS CMS Development Team"
        },
        "license": {
            "name": "Private"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/public/news": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Daftar news publik",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Featured filter",
                        "name": "is_featured",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerPaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.PublicNewsListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/public/news/{slug}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Detail news publik",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PublicNewsDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/public/practice-areas": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Daftar practice area publik",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.PublicPracticeAreaResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/public/teams": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Daftar team publik",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerPaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.PublicTeamListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/public/teams/{slug}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Detail team publik",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SwaggerSuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PublicTeamDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.SwaggerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.PublicNewsDetailResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "featured_image_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_featured": {
                    "type": "boolean"
                },
                "published_at": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.PublicNewsListResponse": {
            "type": "object",
            "properties": {
                "excerpt": {
                    "type": "string"
                },
                "featured_image_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_featured": {
                    "type": "boolean"
                },
                "published_at": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.PublicPracticeAreaResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "display_order": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "dto.PublicTeamDetailResponse": {
            "type": "object",
            "properties": {
                "biography": {
                    "type": "string"
                },
                "degree": {
                    "type": "string"
                },
                "display_order": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "linkedin_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "practice_areas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PublicPracticeAreaResponse"
                    }
                },
                "published_at": {
                    "type": "string"
                },
                "short_description": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "dto.PublicTeamListResponse": {
            "type": "object",
            "properties": {
                "degree": {
                    "type": "string"
                },
                "display_order": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "practice_areas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PublicPracticeAreaResponse"
                    }
                },
                "short_description": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "dto.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.SwaggerPaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Data retrieved successfully"
                },
                "meta": {
                    "$ref": "#/definitions/dto.SwaggerPaginationMeta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.SwaggerPaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 25
                },
                "total_pages": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "dto.SwaggerSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Request successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "XNonce": {
            "description": "Nilai unik untuk mencegah replay attack.",
            "type": "apiKey",
            "name": "X-Nonce",
            "in": "header"
        },
        "XSignature": {
            "description": "HMAC-SHA256 signature dari canonical request.",
            "type": "apiKey",
            "name": "X-Signature",
            "in": "header"
        },
        "XTimestamp": {
            "description": "Unix timestamp dalam detik.",
            "type": "apiKey",
            "name": "X-Timestamp",
            "in": "header"
        }
    }
}